requirements.txt
openai
pyautogen
langchain
faiss-cpu
chromadb
llama-index
python-dotenv
tiktoken